backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
camelot-py==1.0.9
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.116.1
fastapi-cli==0.0.11
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
influxdb3-python==0.15.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.2
pandas-stubs==2.3.2.250827
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
reactivex==4.0.4
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.37.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tabulate==0.9.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
